// src/app/Models/WeightedStats.ts
export class WeightedStats {
    numberOfGamesWeight: number;
    numberOfGoalsWeight: number;
    shotPercentageWeight: number;
    numberOfShotsWeight: number;
    powerPlayGoalsWeight: number;
    goalsPerGameWeight: number;

    opposingGAAWeight: number;
    opposingPKPercWeight: number;
    opposingPKFrequencyWeight: number; 
}